+++ crates/socketioxide-redis/tests/sockets.rs
@@ -49,9 +49,9 @@ fn create_expected_sockets<const N: usize, A: Adapter>(
 pub async fn fetch_sockets() {
     let [io1, io2, io3] = fixture::spawn_servers::<3>();
 
-    io1.ns("/", || ()).await.unwrap();
-    io2.ns("/", || ()).await.unwrap();
-    io3.ns("/", || ()).await.unwrap();
+    io1.ns("/", async || ()).await.unwrap();
+    io2.ns("/", async || ()).await.unwrap();
+    io3.ns("/", async || ()).await.unwrap();
 
     let (_, mut rx1) = io1.new_dummy_sock("/", ()).await;
     let (_, mut rx2) = io2.new_dummy_sock("/", ()).await;
@@ -77,7 +77,8 @@ pub async fn fetch_sockets() {
 #[tokio::test]
 pub async fn fetch_sockets_with_rooms() {
     let [io1, io2, io3] = fixture::spawn_servers::<3>();
-    let handler = |rooms: &'static [&'static str]| move |socket: SocketRef<_>| socket.join(rooms);
+    let handler =
+        |rooms: &'static [&'static str]| async move |socket: SocketRef<_>| socket.join(rooms);
 
     io1.ns("/", handler(&["room1", "room2"])).await.unwrap();
     io2.ns("/", handler(&["room2", "room3"])).await.unwrap();
@@ -106,8 +107,8 @@ pub async fn fetch_sockets_timeout() {
     const TIMEOUT: Duration = Duration::from_millis(50);
     let [io1, io2] = fixture::spawn_buggy_servers(TIMEOUT);
 
-    io1.ns("/", || ()).await.unwrap();
-    io2.ns("/", || ()).await.unwrap();
+    io1.ns("/", async || ()).await.unwrap();
+    io2.ns("/", async || ()).await.unwrap();
 
     let (_, mut rx1) = io1.new_dummy_sock("/", ()).await;
     let (_, mut rx2) = io2.new_dummy_sock("/", ()).await;
@@ -124,8 +125,8 @@ pub async fn fetch_sockets_timeout() {
 pub async fn remote_socket_emit() {
     let [io1, io2] = fixture::spawn_servers();
 
-    io1.ns("/", || ()).await.unwrap();
-    io2.ns("/", || ()).await.unwrap();
+    io1.ns("/", async || ()).await.unwrap();
+    io2.ns("/", async || ()).await.unwrap();
 
     let (_, mut rx1) = io1.new_dummy_sock("/", ()).await;
     let (_, mut rx2) = io2.new_dummy_sock("/", ()).await;
@@ -146,8 +147,8 @@ pub async fn remote_socket_emit() {
 pub async fn remote_socket_emit_with_ack() {
     let [io1, io2] = fixture::spawn_servers();
 
-    io1.ns("/", || ()).await.unwrap();
-    io2.ns("/", || ()).await.unwrap();
+    io1.ns("/", async || ()).await.unwrap();
+    io2.ns("/", async || ()).await.unwrap();
 
     let (_, mut rx1) = io1.new_dummy_sock("/", ()).await;
     let (_, mut rx2) = io2.new_dummy_sock("/", ()).await;