+++ CHANGELOG.md
@@ -4,8 +4,13 @@
 * [parser-common](./crates/parser-common/CHANGELOG.md)
 * [parser-msgpack](./crates/parser-msgpack/CHANGELOG.md)
 * [socketioxide-redis](./crates/socketioxide-redis/CHANGELOG.md)
+* [socketioxide-mongodb](./crates/socketioxide-mongodb/CHANGELOG.md)
 * [socketioxide](./crates/socketioxide/CHANGELOG.md)
 * [engineioxide](./crates/engineioxide/CHANGELOG.md)
+
+# 0.15.2
+## socketioxide
+* fix: issue [#527](https://github.com/Totodore/socketioxide/issues/527). Socketioxide failed to build with the `tracing` flag enable and with `tracing-attributes` set to `v0.1.28`.
 
 # 0.15.1
 ## socketioxide