+++ crates/socketioxide-core/CHANGELOG.md
@@ -1,3 +1,8 @@
+# socketioxide-core 0.17.0
+* feat(*breaking*): remote-adapter packets are now refactored in the core crate. Any adapter implementation can use
+it through the `remote-adapter` flag.
+* MSRV: rust-version is now 1.86 with edition 2024
+
 # socketioxide-core 0.16.1
 * deps: use engineioxide-core 0.1 rather than engineioxide